refactor: condense helpers
