+++ src/client/game-logic/board.ts
@@ -1,6 +1,6 @@
 import { Point } from "../../shared/game";
 import chessData from "./chess-data-import";
-import GamePiece from "./game-piece";
+import GamePiece from "../../shared/game-piece";
 
 export type Grid = {
   name: string;
