+++ src/client/match-logic/base-match.ts
@@ -1,9 +1,8 @@
 import { AbstractMesh } from "@babylonjs/core/Meshes/abstractMesh";
-import { Move, Piece, Point } from "../../shared/game";
+import { Move, Piece, Point, TurnHistory } from "../../shared/game";
 import { Lobby, Player, TEAM } from "../../shared/match";
 import Game from "../game-logic/game";
-import { TurnHistory } from "../game-logic/game-helpers";
-import GamePiece from "../game-logic/game-piece";
+import GamePiece from "../../shared/game-piece";
 import { findByPoint } from "../scenes/scene-helpers";
 
 export interface MatchLogic {
