+++ src/client/match-logic/controller.ts
@@ -3,17 +3,17 @@ import { PickingInfo } from "@babylonjs/core/Collisions/pickingInfo";
 import { IPointerEvent } from "@babylonjs/core/Events/deviceInputEvents";
 import { AbstractMesh } from "@babylonjs/core/Meshes/abstractMesh";
 import type { Nullable } from "@babylonjs/core/types";
-import { Point } from "../../shared/game";
+import { Point, TurnHistory } from "../../shared/game";
 import { ControllerOptions, LOBBY_TYPE } from "../../shared/match";
 import { Message } from "../components/modals/message-modal";
-import { doMovesMatch, TurnHistory } from "../game-logic/game-helpers";
-import GamePiece from "../game-logic/game-piece";
+import GamePiece from "../../shared/game-piece";
 import { rotateCamera } from "../scenes/animation/camera";
 import calcTurnAnimation from "../scenes/animation/turn-animation";
 import { displayPieceMoves, findByPoint } from "../scenes/scene-helpers";
 import { GameScene, SceneManager, Scenes } from "../scenes/scene-manager";
 import { LocalMatch } from "./local-match";
 import { OnlineMatch } from "./online-match";
+import { doMovesMatch } from "../game-logic/helpers";
 
 export class Controller {
   sceneManager: SceneManager;
@@ -48,7 +48,8 @@ export class Controller {
     const gameScene = this.sceneManager.switchScene(Scenes.GAME);
     if (gameScene) {
       this.subscribeGameInput(gameScene);
-      this.resetView();
+      this.updateMeshesRender();
+      this.resetCamera();
     }
   }
