+++ src/client/match-logic/local-match.ts
@@ -1,6 +1,6 @@
 import { Point } from "../../shared/game";
 import { LOBBY_TYPE, Lobby, Player } from "../../shared/match";
-import GamePiece from "../game-logic/game-piece";
+import GamePiece from "../../shared/game-piece";
 import { BaseMatch, MatchLogic } from "./base-match";
 
 export class LocalMatch extends BaseMatch implements MatchLogic {
