+++ src/client/scenes/scene-helpers.ts
@@ -3,7 +3,7 @@ import { Vector3 } from "@babylonjs/core/Maths/math.vector.js";
 import { MeshBuilder } from "@babylonjs/core/Meshes/meshBuilder";
 import { GameScene } from "./scene-manager";
 import { pointIndexMap, pointPositionMap } from "../data/point-map-data";
-import GamePiece from "../game-logic/game-piece";
+import GamePiece from "../../shared/game-piece";
 import { Scene } from "@babylonjs/core/scene";
 import { Move, Point } from "../../shared/game";
